+++ js/screens/artist-screen.js
@@ -1,58 +1,22 @@
-// Экран выбора исполнителя
-import {renderElement} from '../utils.js';
-import {levels} from '../data';
+import ArtistView from '../views/artist-view.js';
+import {initAutoplay, addArtistListener} from '../audio/audio.js';
+import {INITIAL_STATE} from '../data';
 import changeScreen from '../change-screen.js';
 import {changeLives} from '../game.js';
-import {initAutoplay, addArtistListener} from '../audio/audio.js';
-
-const artistTemplate = (level) => `<div><div class="game__track">
-        <button class="track__button track__button--play" type="button"></button>
-        <audio src="${level.task.src}"></audio>
-      </div>
-      <form class="game__artist">
-        ${level.answers.map((answer, i) => `<div class="artist">
-          <input class="artist__input visually-hidden" type="radio" name="answer" value="artist-${i}" id="answer-${i}">
-          <label class="artist__name" for="answer-${i}">
-            <img class="artist__picture" src="${answer.image}" alt="${answer.artist}">
-            ${answer.artist}
-          </label>
-        </div>`).join(``)}
-      </form></div>`;
 
 const artistScreen = (state) => {
+  const screen = new ArtistView(state);
 
-  // Текущий уровень
-  const currentLevel = levels[state.level];
-
-  // Рендерим экран
-  const artistElement = renderElement(artistTemplate(currentLevel));
-
-  // Кнопка Play и трек
-  const playButton = artistElement.querySelector(`.track__button`);
-  const audio = artistElement.querySelector(`audio`);
-
-  addArtistListener(playButton, audio);
-
-  // Меняем вид кнопки Play
-  initAutoplay(audio, playButton);
-
-  // Массив элементов артистов
-  const artists = artistElement.querySelectorAll(`.artist`);
-
-  // Обрабатываем клик
-  artists.forEach((artist) => {
-    artist.addEventListener(`click`, () => {
-      if (artist.querySelector(`img`).src === currentLevel.task.image) {
-        state.answers.push({answer: true, time: 30});
-        changeScreen(state);
-      } else {
-        state.answers.push({answer: false, time: 30});
-        changeScreen(changeLives(state, state.lives - 1));
-      }
-    });
-  });
-
-  return artistElement;
+  screen.onAnswer = (item, answer) => {
+    if (item.querySelector(`img`).src === answer) {
+      state.answers.push({answer: true, time: 30});
+      changeScreen(state);
+    } else {
+      state.answers.push({answer: false, time: 30});
+      changeScreen(changeLives(state, state.lives - 1));
+    }
+  };
+  return screen;
 };
 
 export default artistScreen;
