+++ js/screens/fail-tries-screen.js
@@ -1,25 +1,13 @@
-// Экран поражения при завершении количества попыток
-import {renderElement} from '../utils.js';
-import backButtonClickHandler from './back-button-handler.js';
+import {INITIAL_STATE} from '../data';
+import FailTriesView from '../views/fail-tries-view.js';
+import {renderScreen} from '../utils.js';
 
-const failTriesTemplate = () => `<section class="result">
-  <div class="result__logo">
-    <img src="img/melody-logo.png" alt="Угадай мелодию" width="186" height="83">
-  </div>
-  <h2 class="result__title">Какая жалость!</h2>
-  <p class="result__total result__total--fail">У вас закончились все попытки. Ничего, повезёт в следующий раз!</p>
-  <button class="result__replay" type="button">Попробовать ещё раз</button>
-</section>`;
+const failTriesScreen = (state) => {
+  const screen = new FailTriesView(state);
 
-const failTriesScreen = () => {
-  const screen = renderElement(failTriesTemplate());
-
-  // Кнопка "Сыграть ещё раз"
-  const replayButton = screen.querySelector(`.result__replay`);
-
-  // Открываем приветственный экран при клике
-  replayButton.addEventListener(`click`, backButtonClickHandler);
-
+  screen.onReplayButton = () => {
+    renderScreen(welcomeScreen(INITIAL_STATE).element);
+  };
   return screen;
 };
