+++ js/views/fail-time-view.js
@@ -1,0 +1,28 @@
+import AbstractView from './abstract-view.js';
+
+export default class FailTimeView extends AbstractView {
+  constructor(state) {
+    super();
+    this.state = state;
+  }
+
+  get template() {
+    return `<section class="result">
+  <div class="result__logo">
+    <img src="img/melody-logo.png" alt="Угадай мелодию" width="186" height="83">
+  </div>
+  <h2 class="result__title">Увы и ах!</h2>
+  <p class="result__total result__total--fail">Время вышло! Вы не успели отгадать все мелодии</p>
+  <button class="result__replay" type="button">Попробовать ещё раз</button>
+</section>`;
+  }
+
+  onReplayButton() {}
+
+  bind() {
+    this.element.querySelector(`.result__replay`).addEventListener(`click`, (evt) => {
+      evt.preventDefault();
+      this.onReplayButton(this.state);
+    });
+  }
+}
